Clean PR

+++ crates/blockchain/payload.rs
@@ -376,7 +376,7 @@ impl Blockchain {
         let max_blob_number_per_block = chain_config
             .get_fork_blob_schedule(context.payload.header.timestamp)
             .map(|schedule| schedule.max)
-            .unwrap_or_default();
+            .unwrap_or_default() as usize;
 
         debug!("Fetching transactions from mempool");
         // Fetch mempool transactions
@@ -388,8 +388,7 @@ impl Blockchain {
                 debug!("No more gas to run transactions");
                 break;
             };
-            if !blob_txs.is_empty()
-                && context.blobs_bundle.blobs.len() >= max_blob_number_per_block as usize
+            if !blob_txs.is_empty() && context.blobs_bundle.blobs.len() >= max_blob_number_per_block
             {
                 debug!("No more blob gas to run blob transactions");
                 blob_txs.clear();
@@ -482,16 +481,14 @@ impl Blockchain {
         let max_blob_number_per_block = chain_config
             .get_fork_blob_schedule(context.payload.header.timestamp)
             .map(|schedule| schedule.max)
-            .unwrap_or_default();
+            .unwrap_or_default() as usize;
         let Some(blobs_bundle) = self.mempool.get_blobs_bundle(tx_hash)? else {
             // No blob tx should enter the mempool without its blobs bundle so this is an internal error
             return Err(
                 StoreError::Custom(format!("No blobs bundle found for blob tx {tx_hash}")).into(),
             );
         };
-        if context.blobs_bundle.blobs.len() + blobs_bundle.blobs.len()
-            > max_blob_number_per_block as usize
-        {
+        if context.blobs_bundle.blobs.len() + blobs_bundle.blobs.len() > max_blob_number_per_block {
             // This error will only be used for debug tracing
             return Err(EvmError::Custom("max data blobs reached".to_string()).into());
         };
